requirements.txt
biopython
pandas
plotly # Or matplotlib
# For command-line argument parsing
argparse